test(comments): break apart inline comment range tests
Have each test in it's own file to make them easier to write and read.
Ditch snapshot testing and test against hard coded values that makes
it easier to undstand what is going on.
